moving processWorkfflow to WorkFlowService
