+++ zieta/src/main/java/com/zietaproj/zieta/service/WorkFlowRequestService.java
@@ -17,5 +17,7 @@ public interface WorkFlowRequestService {
 	public List<WorkFlowRequestorData> findByRequestorId(long requestorId);
 
 	public WorkflowRequest findByTsIdAndApproverId(long tsId, long approverId);
+	
+	public void processWorkFlow(Long id, short actionType, String comments);
 
 }
