+++ zieta/src/main/java/com/zietaproj/zieta/serviceImpl/TimeSheetServiceImpl.java
@@ -254,84 +254,6 @@ public class TimeSheetServiceImpl implements TimeSheetService {
 	}
 	
 	
-	@Transactional
-	public void processWorkFlow(Long workFlowRequestId, short actionType, String comments) {
-		
-		WorkflowRequest workFlowRequest = workflowRequestRepository.findById(workFlowRequestId).get();
-		
-		List<ProcessSteps> processStepsList = processStepsRepository.findByClientIdAndProjectIdAndProjectTaskId(
-				workFlowRequest.getClientId(),workFlowRequest.getProjectId(), workFlowRequest.getProjectTaskId());
-		int workFlowDepth = processStepsList.size();
-		
-		TSInfo tsInfo = tSInfoRepository.findById(workFlowRequest.getTsId()).get();
-		
-		List<WorkflowRequest> workFlowRequestList = workflowRequestRepository.findByTsId(workFlowRequest.getTsId());
-		Map<String, Long> statusTypes =  getStatus(tsInfo);
-		
-		for (int i = 0; workFlowRequestList != null && i < workFlowRequestList.size(); i++) {
-
-			workFlowInAction(actionType, comments, workFlowRequestList.get(i), processStepsList, workFlowDepth, tsInfo,
-					statusTypes);
-			if (actionType == ActionType.APPROVE.getActionType() && workFlowRequestList.get(i).getId() != workFlowRequestId) {
-				
-				workflowRequestRepository.deleteById(workFlowRequestList.get(i).getId());
-			}
-		}
-		
-	}
-
-
-	private void workFlowInAction(short actionType, String comments, WorkflowRequest workFlowRequest,
-			List<ProcessSteps> processStepsList, int workFlowDepth, TSInfo tsInfo, Map<String, Long> statusTypes) {
-		if (actionType == ActionType.APPROVE.getActionType()) {
-			// promote the approval to next level
-			long currentStep = workFlowRequest.getCurrentStep();
-			if (currentStep <= workFlowDepth) {
-				if (currentStep != workFlowDepth) {
-					long nextStep = currentStep + 1;
-					List<ProcessSteps> processStepFiltered = processStepsList.stream()
-							.filter(step -> step.getStepId().equals(nextStep)).collect(Collectors.toList());
-					workFlowRequest.setCurrentStep(nextStep);
-					// Promoting to next level approval
-					workFlowRequest.setComments(comments);
-					workFlowRequest.setApproverId(Long.valueOf(processStepFiltered.get(0).getApproverId()));
-					workFlowRequest.setStateType(StateType.INPROCESS.getStateTypeId());
-
-					// approved from the current step point of view
-					workFlowRequest.setActionType(ActionType.APPROVE.getActionType());
-					workFlowRequest.setActionDate(new Date());
-				} else {
-					// Final Approval Done
-					workFlowRequest.setComments(comments);
-					tsInfo.setStatusId(statusTypes.get(Status.APPROVED.getStatus()));
-					tSInfoRepository.save(tsInfo);
-					workFlowRequest.setStateType(StateType.COMPLETE.getStateTypeId());
-					workFlowRequest.setActionType(ActionType.APPROVE.getActionType());
-					workFlowRequest.setActionDate(new Date());
-				}
-
-			}
-		} else if (actionType == ActionType.REJECT.getActionType()) {
-			// Request Rejected
-			workFlowRequest.setComments(comments);
-			workFlowRequest.setStateType(StateType.REJECT.getStateTypeId());
-			workFlowRequest.setActionType(ActionType.REJECT.getActionType());
-			tsInfo.setStatusId(statusTypes.get(Status.REJECTED.getStatus()));
-			tSInfoRepository.saveAndFlush(tsInfo);
-			workFlowRequest.setActionDate(new Date());
-		} else if (actionType == ActionType.REVISE.getActionType()) {
-			// Request sent for revise
-			workFlowRequest.setComments(comments);
-			workFlowRequest.setActionDate(new Date());
-			workFlowRequest.setStateType(StateType.INITIAL.getStateTypeId());
-			workFlowRequest.setActionType(ActionType.REVISE.getActionType());
-			tsInfo.setStatusId((statusTypes.get(Status.DRAFT.getStatus())));
-			tSInfoRepository.save(tsInfo);
-			// Do we need to create the new Request or if its with the same request , need
-			// to start with Initial Step ?
-		}
-	}
-	
 	@Override
 	public List<TimeEntriesByTimesheetIDResponse> getTimeEntriesByTsID(Long tsId) {
 		short notDeleted=0;
